100 - Move JSONDataSource to separate file
