+++ tests/tel2puml/otel_to_pv/data_sources/json_data_source/test_json_datasource.py
@@ -1,16 +1,15 @@
 """Tests for otel_data_source module."""
-
 import json
 import os
-import yaml
-import pytest
 import shutil
 
+import yaml
+import pytest
 from pytest import FixtureRequest
 from pathlib import Path
 from unittest.mock import mock_open, patch
 
-from tel2puml.otel_to_pv.data_sources.base import (
+from tel2puml.otel_to_pv.data_sources.json_data_source.json_datasource import (
     JSONDataSource,
 )
 from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
