+++ tests/tel2puml/otel_to_pv/test_ingest_data.py
@@ -17,7 +17,7 @@ from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
     IngestTypes
 )
 from tel2puml.otel_to_pv.data_holders import SQLDataHolder
-from tel2puml.otel_to_pv.data_sources.base import (
+from tel2puml.otel_to_pv.data_sources import (
     JSONDataSource,
 )
